@@ -21,18 +21,18 @@
     "clean-full:package_modules": "rimraf ./packages/*/node_modules",
     "compile": "node ./scripts/build.js",
     "copy": "cd packages/wdio-cli && npm run copy && cd ../../",
-    "generate:typings": "ts-node-esm ./scripts/type-generation/generate-all-typings.ts",
+    "generate:typings": "node --loader ts-node/esm ./scripts/type-generation/generate-all-typings.ts",
     "link": "node ./scripts/link.js",
     "pushReleaseTag": "node ./scripts/pushTags.js",
     "setup-full": "run-s bootstrap-full build",
     "setup": "run-s bootstrap build",
     "docs": "run-s docs:*",
     "docs:setup": "cd ./website && npm install && cd ..",
-    "docs:generate": "npm run clean:website && ts-node-esm ./scripts/docs-generation/generateDocs.ts",
+    "docs:generate": "npm run clean:website && node --loader ts-node/esm ./scripts/docs-generation/generateDocs.ts",
     "docs:build": "cd ./website && npm run build && cd ..",
     "docs:deploy": "node ./scripts/updateDocs.js",
     "test": "run-s test:eslint test:ejslint test:typings test:unit:run test:smoke test:component",
-    "test:depcheck": "ts-node-esm ./scripts/depcheck.ts",
+    "test:depcheck": "node --loader ts-node/esm ./scripts/depcheck.ts",
     "test:eslint": "eslint --cache packages examples scripts tests",
     "test:ejslint": "ejslint packages/wdio-cli/src/templates/**/*",
     "test:typings": "run-s test:typings:*",
@@ -66,10 +66,10 @@
     "postinstall:vulnerabilities": "node ./.husky/postinstall.js"
   },
   "devDependencies": {
-    "@cucumber/cucumber": "^9.1.0",
+    "@cucumber/cucumber": "^9.1.2",
     "@octokit/rest": "^19.0.7",
     "@preact/preset-vite": "^2.5.0",
-    "@sveltejs/vite-plugin-svelte": "^2.1.1",
+    "@sveltejs/vite-plugin-svelte": "^2.2.0",
     "@tsconfig/node18": "^2.0.0",
     "@types/archiver": "^5.3.2",
     "@types/cheerio": "^0.22.31",
@@ -78,34 +78,34 @@
     "@types/inquirer": "^9.0.3",
     "@types/lodash.clonedeep": "^4.5.7",
     "@types/lodash.zip": "^4.2.7",
-    "@types/node": "^18.16.1",
+    "@types/node": "^20.1.1",
     "@types/shelljs": "^0.8.12",
     "@types/split2": "^4.2.0",
     "@types/uuid": "^9.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
-    "@typescript-eslint/utils": "^5.59.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.5",
+    "@typescript-eslint/parser": "^5.59.5",
+    "@typescript-eslint/utils": "^5.59.5",
     "@vitejs/plugin-react": "^4.0.0",
-    "@vitejs/plugin-vue": "^4.2.0",
+    "@vitejs/plugin-vue": "^4.2.1",
     "@vitest/coverage-c8": "^0.30.1",
-    "aws-sdk": "^2.1365.0",
+    "aws-sdk": "^2.1374.0",
     "c8": "^7.13.0",
     "camelcase": "^7.0.1",
     "chalk": "^5.2.0",
     "cheerio": "^1.0.0-rc.12",
-    "chromedriver": "^112.0.0",
+    "chromedriver": "^113.0.0",
     "copyfiles": "^2.4.1",
     "depcheck": "^1.4.3",
     "ejs": "^3.1.9",
     "ejs-lint": "^2.0.0",
-    "eslint": "^8.39.0",
+    "eslint": "^8.40.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-unicorn": "^46.0.0",
-    "eslint-plugin-wdio": "^8.0.14",
+    "eslint-plugin-unicorn": "^47.0.0",
+    "eslint-plugin-wdio": "^8.8.7",
     "husky": "^8.0.3",
-    "inquirer": "^9.2.0",
-    "jsdom": "^21.1.1",
-    "lerna": "^6.6.1",
+    "inquirer": "^9.2.2",
+    "jsdom": "^22.0.0",
+    "lerna": "^6.6.2",
     "lerna-changelog": "^2.2.0",
     "markdox": "^0.1.10",
     "mime-types": "^2.1.35",
