@@ -6,12 +6,26 @@ import type { BeforeCommandArgs, AfterCommandArgs } from '@wdio/reporter'
 import { v4 as uuidv4 } from 'uuid'
 import type { Pickle, ITestCaseHookParameter } from './cucumber-types.js'
 
-import { frameworkSupportsHook, getCloudProvider, getGitMetaData, getHookType, getScenarioExamples, getUniqueIdentifier, getUniqueIdentifierForCucumber, isBrowserstackSession, isScreenshotCommand, removeAnsiColors, sleep, uploadEventData } from './util.js'
+import {
+    frameworkSupportsHook,
+    getCloudProvider,
+    getGitMetaData,
+    getHookType,
+    getScenarioExamples,
+    getUniqueIdentifier,
+    getUniqueIdentifierForCucumber,
+    isBrowserstackSession,
+    isScreenshotCommand,
+    o11yClassErrorHandler,
+    removeAnsiColors,
+    sleep,
+    uploadEventData
+} from './util.js'
 import type { TestData, TestMeta, PlatformMeta, UploadType } from './types.js'
 import RequestQueueHandler from './request-handler.js'
 import { DATA_SCREENSHOT_ENDPOINT, DEFAULT_WAIT_INTERVAL_FOR_PENDING_UPLOADS, DEFAULT_WAIT_TIMEOUT_FOR_PENDING_UPLOADS } from './constants.js'
 
-export default class InsightsHandler {
+class _InsightsHandler {
     private _tests: Record<string, TestMeta> = {}
     private _hooks: Record<string, string[]> = {}
     private _platformMeta: PlatformMeta
@@ -513,3 +527,10 @@ export default class InsightsHandler {
         return getUniqueIdentifier(test, this._framework)
     }
 }
+
+// https://github.com/microsoft/TypeScript/issues/6543
+const InsightsHandler: typeof _InsightsHandler = o11yClassErrorHandler(_InsightsHandler)
+type InsightsHandler = _InsightsHandler
+
+export default InsightsHandler
+
